requirements.txt
opencv-python
streamlit
Pillow
